use job to process file resources move updating to public id to VSCodeIdService
