+++ server/src/main/java/org/eclipse/openvsx/publish/PublishExtensionVersionJobRequest.java
@@ -1,0 +1,37 @@
+/** ******************************************************************************
+ * Copyright (c) 2023 Precies. Software Ltd and others
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License v. 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0.
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ * ****************************************************************************** */
+package org.eclipse.openvsx.publish;
+
+import org.jobrunr.jobs.lambdas.JobRequest;
+import org.jobrunr.jobs.lambdas.JobRequestHandler;
+
+public class PublishExtensionVersionJobRequest implements JobRequest {
+
+    private long downloadId;
+
+    public PublishExtensionVersionJobRequest() {}
+
+    public PublishExtensionVersionJobRequest(long downloadId) {
+        this.downloadId = downloadId;
+    }
+
+    @Override
+    public Class<? extends JobRequestHandler> getJobRequestHandler() {
+        return PublishExtensionVersionJobRequestHandler.class;
+    }
+
+    public long getDownloadId() {
+        return downloadId;
+    }
+
+    public void setDownloadId(long downloadId) {
+        this.downloadId = downloadId;
+    }
+}
